Add AccountHistory feature with related entities and mappers
Implemented the AccountHistory functionality including the mapper, handler, repository, and new specifications for date filtering. Integrated the feature into transaction handling to ensure account history records operations.

+++ src/main/java/br/dev/jstec/tyginvestiment/dto/AssetTransactionDto.java
@@ -15,7 +15,7 @@ public class AssetTransactionDto {
     private String description;
     private String transactionType;
     private LocalDate transactionDate;
-    private Double quantity;
+    private BigDecimal quantity;
     private BigDecimal value;
     private String assetId;
     private Long accountId;
